fix that if department was deleted in all students and teachers in this department field department = null

+++ src/main/java/ui/DepartmentMenu.java
@@ -235,9 +235,14 @@ public class DepartmentMenu {
             Optional<Department> maybeDepartment = DepartmentService.findById(id);
             if (maybeDepartment.isPresent()) {
                 Department foundDepartment = maybeDepartment.get();
-                DepartmentService.delete(foundDepartment);
-                System.out.println("\nВи успішно видалили кафедру. Оновлені дані: ");
-                printAll(DepartmentService.getAll());
+                System.out.println("\nУВАГА!!!");
+                System.out.println("Після видалення кафедри всі студенти та викладачі, що приписані до цього факультету, більше не будуть до нього приписані.");
+                int makingSure = InputReader.readInt("Введіть 1, якщо точно хочете видалити, або 0, щоби відмінити дію: ", 0, 1);
+                if(makingSure==1){
+                    DepartmentService.delete(foundDepartment);
+                    System.out.println("\nВи успішно видалили кафедру " + foundDepartment.getName() + ". Оновлені дані: ");
+                    printAll(DepartmentService.getAll());
+                }
             }else{
                 System.out.println("\nКафедру з унікальним ідентифікатором " + id + " не знайдено.");
             }
